add: specialOffer class-controller-model ecc...

+++ src/main/java/com/pizzeria/repository/PizzaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.pizzeria.model.Pizza;
 
 @Repository
-public interface PizzaRepository extends JpaRepository<Pizza, Integer> {
+public interface PizzaRepository extends JpaRepository<Pizza, Long> {
 
 	/* QUEARY EQUIVALENTE
 	 * 
@@ -20,3 +20,4 @@ public interface PizzaRepository extends JpaRepository<Pizza, Integer> {
 	public List<Pizza> findByName(String name);
 
 }
+
